moved original project to archive and created new gem
