+++ Rakefile
@@ -1,6 +1,6 @@
-desc "Run tests"
-task :test do
-  Dir["#{__dir__}/spec/**/*_spec.rb"].each(&method(:load))
-end
+require "bundler/gem_tasks"
+require "rspec/core/rake_task"
 
-task :default => :test
+RSpec::Core::RakeTask.new(:spec)
+
+task :default => :spec
