+++ archive/Rakefile
@@ -1,0 +1,6 @@
+desc "Run tests"
+task :test do
+  Dir["#{__dir__}/spec/**/*_spec.rb"].each(&method(:load))
+end
+
+task :default => :test
