+++ lib/vine_dl/version.rb
@@ -1,0 +1,3 @@
+module VineDl
+  VERSION = "0.1.0"
+end
